Switching to grid-flow-col-dense
Solving the tiling problem is unnecessarily hard.

Instead I'm going to use the grid-flow-col-dense option in CSS and pipe each tile into a randomly selected type of box (to get variation in size). This may leave some blanks at the end but I think I'm okay with it for the time being.

This may require adding more properties to the object that is piped into the tile so that the title lengths etc are of appropriate length.