Better margins handling, a new PaperAccordion component started
The big new addition is the creation of a paper "accordion" component.

The component has some basic expansion functionality with animation of some of the underlying meta-data with pretty effects/saving space.

The component still needs some styling to pretty up the spacing and justification of  a lot of the flex items:

Component still needs to include the following pieces of information:
--- Status (for working papers)
---- Journal (for published)
---- Buttons
------ Explanation (blog post)
------- Download
-------- Data/GitHub Replication materials
-------- "In the press"