Attempting blog *post* pages
Very tentative attempt at getting MDX files as blog post pages. Not working very well...

So far...
I can pull metadata to generate an index page but pulling the actual body of the .MDX file is a challenge. Next.JS' dynamic routing and SSG isn't playing nicely with the current folder structure. Because browser doesn't have access to the file system (module), I'm struggling to pass through the contents of an individual page to a "post"-type page.

Going to try a different folder structure that follows the wishes of NextJS more closely.